+++ src/SemanticJob_Op.cpp
@@ -660,7 +660,7 @@ bool SemanticJob::resolveUserOp(SemanticContext* context, const Utf8& name, cons
         context->result = ContextResult::Done;
     }
 
-    uint32_t castFlags = CASTFLAG_UNCONST | CASTFLAG_AUTO_OPCAST | CASTFLAG_UFCS | CASTFLAG_ACCEPT_PENDING | CASTFLAG_PARAMS;
+    uint64_t castFlags = CASTFLAG_UNCONST | CASTFLAG_AUTO_OPCAST | CASTFLAG_UFCS | CASTFLAG_ACCEPT_PENDING | CASTFLAG_PARAMS;
     if (justCheck)
         castFlags |= CASTFLAG_JUST_CHECK;
     if (job->cacheMatches.empty())
