+++ src/TypeInfo.h
@@ -123,7 +123,7 @@ struct BadSignatureInfos
     ComputedValue* badGenValue2              = nullptr;
     MatchResult    matchResult               = MatchResult::Ok;
 
-    uint32_t castErrorFlags           = 0;
+    uint64_t castErrorFlags           = 0;
     int      badSignatureParameterIdx = 0;
     int      badSignatureNum1         = 0;
     int      badSignatureNum2         = 0;
@@ -291,7 +291,7 @@ struct TypeInfo
     bool isFakeAlias()                      { return (flags & TYPEINFO_FAKE_ALIAS); }
     // clang-format on
 
-    virtual bool        isSame(TypeInfo* from, uint32_t castFlags);
+    virtual bool        isSame(TypeInfo* from, uint64_t castFlags);
     virtual TypeInfo*   clone() = 0;
     virtual int         numRegisters();
     virtual Utf8        getDisplayName();
@@ -338,7 +338,7 @@ struct TypeInfo
 struct TypeInfoParam
 {
     int            numRegisters();
-    bool           isSame(TypeInfoParam* to, uint32_t castFlags);
+    bool           isSame(TypeInfoParam* to, uint64_t castFlags);
     TypeInfoParam* clone();
     void           allocateComputedValue();
 
@@ -372,7 +372,7 @@ struct TypeInfoNative : public TypeInfo
         valueInteger = 0;
     }
 
-    bool      isSame(TypeInfo* from, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* from, uint64_t castFlags) override;
     TypeInfo* clone() override;
 
     union
@@ -401,7 +401,7 @@ struct TypeInfoEnum : public TypeInfo
     {
     }
 
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
     bool      contains(const Utf8& valueName);
 
@@ -421,9 +421,9 @@ struct TypeInfoFuncAttr : public TypeInfo
 
     TypeInfo* clone() override;
     void      computeWhateverName(Utf8& resName, uint32_t nameType) override;
-    bool      isSame(TypeInfo* from, uint32_t castFlags) override;
-    bool      isSame(TypeInfoFuncAttr* other, uint32_t castFlags, BadSignatureInfos& bi);
-    bool      isSame(TypeInfoFuncAttr* from, uint32_t castFlags);
+    bool      isSame(TypeInfo* from, uint64_t castFlags) override;
+    bool      isSame(TypeInfoFuncAttr* other, uint64_t castFlags, BadSignatureInfos& bi);
+    bool      isSame(TypeInfoFuncAttr* from, uint64_t castFlags);
     void      match(SymbolMatchContext& context);
     bool      returnByCopy();
     bool      returnByValue();
@@ -459,7 +459,7 @@ struct TypeInfoPointer : public TypeInfo
     }
 
     void      computeWhateverName(Utf8& resName, uint32_t nameType) override;
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
 
     TypeInfo* pointedType = nullptr;
@@ -479,7 +479,7 @@ struct TypeInfoArray : public TypeInfo
     }
 
     void      computeWhateverName(Utf8& resName, uint32_t nameType) override;
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
 
     TypeInfo* pointedType = nullptr;
@@ -499,7 +499,7 @@ struct TypeInfoSlice : public TypeInfo
     }
 
     void      computeWhateverName(Utf8& resName, uint32_t nameType) override;
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
 
     TypeInfo* pointedType = nullptr;
@@ -517,7 +517,7 @@ struct TypeInfoList : public TypeInfo
         return 1;
     }
 
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
     void      computeWhateverName(Utf8& resName, uint32_t nameType) override;
     Utf8      computeTupleName(JobContext* context);
@@ -534,7 +534,7 @@ struct TypeInfoVariadic : public TypeInfo
     {
     }
 
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
     void      computeWhateverName(Utf8& resName, uint32_t nameType) override;
 
@@ -549,7 +549,7 @@ struct TypeInfoGeneric : public TypeInfo
         flags |= TYPEINFO_GENERIC;
     }
 
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
 
     TypeInfo* rawType = nullptr;
@@ -569,7 +569,7 @@ struct TypeInfoStruct : public TypeInfo
         return 1;
     }
 
-    bool           isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool           isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo*      clone() override;
     void           computeWhateverName(Utf8& resName, uint32_t nameType) override;
     void           match(SymbolMatchContext& context);
@@ -622,7 +622,7 @@ struct TypeInfoAlias : public TypeInfo
     }
 
     void      computeWhateverName(Utf8& resName, uint32_t nameType) override;
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
 
     int numRegisters() override
@@ -642,7 +642,7 @@ struct TypeInfoCode : public TypeInfo
     {
     }
 
-    bool      isSame(TypeInfo* to, uint32_t castFlags) override;
+    bool      isSame(TypeInfo* to, uint64_t castFlags) override;
     TypeInfo* clone() override;
 
     AstNode* content = nullptr;
